Reads the VeryLongStringRecord info record
