+++ SpssLib/FileParser/Records/BaseInfoRecord.cs
