+++ SpssLib/FileParser/Records/InfoRecordParser.cs
