+++ SpssLib/FileParser/Records/VeryLongStringRecord.cs
@@ -1,47 +1,31 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
+using System.IO;
 using System.Text;
 
 namespace SpssLib.FileParser.Records
 {
-    public class VeryLongStringRecord
+    public class VeryLongStringRecord : VariableDataInfoRecord<int>
     {
-        private InfoRecord record;
+        public override int SubType { get { return InfoRecordType.VeryLongString; } }
 
-        internal VeryLongStringRecord(InfoRecord record)
+        protected override int DecodeValue(string stringValue)
         {
-            if (record.SubType != 14 || record.ItemSize != 1)
-                throw new UnexpectedFileFormatException();
-            this.record = record;
-
-            this.LongStringDictionary = new Dictionary<string, string>();
-
-            // Not very efficient, but for now the best way I can come up with
-            //   without sacrificing the Inforecords-design.
-            var originalBytes = this.record.Items.Where(item => item[0] != 0).Select(item => item[0]).ToArray();
-            var dictionaryString = Encoding.ASCII.GetString(originalBytes);
-
-            // split on tabs:
-            var entries = dictionaryString.Split('\t');
-
-            foreach (var entry in entries)
-            {
-                if (!string.IsNullOrEmpty(entry.Trim()))
-                {
-                    var values = entry.Split('=');
-                    if (values.Length >= 2)
-                    {
-                        this.LongStringDictionary.Add(values[0], values[1]);
-                    }
-                }
-            }
-
+            // TODO exception if fails
+            int lenght;
+            int.TryParse(stringValue, out lenght);
+            return lenght;
         }
-        public Dictionary<string, string> LongStringDictionary
+
+        protected override string EncodeValue(int value)
         {
-            get;
-            private set;
+            throw new NotImplementedException();
+        }
+
+        public override void RegisterMetadata(MetaData metaData)
+        {
+            metaData.VeryLongStrings = this;
+            Metadata = metaData;
         }
     }
 }
